Added Device adding and Details Screen
added Device adding screen with api
added Details screen

+++ MiTestAnirban/app/src/main/java/com/mi/mitestanirban/MainActivity.java
@@ -1,9 +1,9 @@
 package com.mi.mitestanirban;
 
+import android.content.Intent;
 import android.database.Cursor;
 import android.os.Bundle;
 import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
 import android.support.v7.app.AlertDialog;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.LinearLayoutManager;
@@ -11,14 +11,20 @@ import android.support.v7.widget.RecyclerView;
 import android.support.v7.widget.Toolbar;
 import android.view.View;
 import android.widget.LinearLayout;
+import android.widget.TextView;
 import android.widget.Toast;
 
+import com.mi.mitestanirban.events.GetAndroidVersionListEvent;
 import com.mi.mitestanirban.events.GetDeviceListEvent;
-import com.mi.mitestanirban.model.Devices;
-import com.mi.mitestanirban.model.GetAllDeviceListJob;
+import com.mi.mitestanirban.events.PassDeviceEvent;
+import com.mi.mitestanirban.jobs.GetAllAndroidVersionListJob;
+import com.mi.mitestanirban.jobs.GetAllDeviceListJob;
+import com.mi.mitestanirban.model.AndroidDeviceWithVersion;
+import com.mi.mitestanirban.model.AndroidVersion;
+import com.mi.mitestanirban.model.Device;
 import com.mi.mitestanirban.utils.MyDatabase;
 import com.mi.mitestanirban.utils.ReusableClass;
-import com.mi.mitestanirban.widgets.DeviceListAdapter;
+import com.mi.mitestanirban.widgets.DeviceWithAndroidVersionListAdapter;
 
 import java.util.ArrayList;
 
@@ -36,10 +42,12 @@ public class MainActivity extends AppCompatActivity {
     RecyclerView deviceList;
     @Bind(R.id.progress_container)
     LinearLayout progressContainer;
+    @Bind(R.id.textViewNoData)
+    TextView textViewNoData;
 
-    private DeviceListAdapter mAdapter;
+    private DeviceWithAndroidVersionListAdapter mAdapter;
     private MyDatabase db;
-
+    private boolean noData = false;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -51,21 +59,28 @@ public class MainActivity extends AppCompatActivity {
         fab.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                Snackbar.make(view, "Replace with your own action", Snackbar.LENGTH_LONG)
-                        .setAction("Action", null).show();
+                Intent i = new Intent(MainActivity.this, AddingDeviceActivity.class);
+                startActivity(i);
             }
         });
 
         db = new MyDatabase(this);
-        mAdapter = new DeviceListAdapter(this);
+        mAdapter = new DeviceWithAndroidVersionListAdapter(this);
 
         LinearLayoutManager mLayoutManager = new LinearLayoutManager(this);
         deviceList.setLayoutManager(mLayoutManager);
         deviceList.setAdapter(mAdapter);
 
+        Cursor cursor = db.getAllDeviceWithAndroidVersion();
+        if (cursor.getCount() > 0) {
+            noData = false;
+            addingDataToList(cursor);
+        } else
+            noData = true;
 
         if (ReusableClass.haveNetworkConnection(this)) {
-            setListShown(false, true);
+            if (noData)
+                setListShown(false, true);
             MyApplication.addJobInBackground(new GetAllDeviceListJob());
         } else
             Toast.makeText(this, R.string.error_internet_connection, Toast.LENGTH_LONG).show();
@@ -78,7 +93,7 @@ public class MainActivity extends AppCompatActivity {
     @Override
     public void onStart() {
         super.onStart();
-        EventBus.getDefault().register(this);
+        EventBus.getDefault().registerSticky(this);
     }
 
     @Override
@@ -90,32 +105,22 @@ public class MainActivity extends AppCompatActivity {
     public void onEventMainThread(GetDeviceListEvent.Success event) {
         setListShown(true, true);
 
-        db.deleteAllDevice();
+        if (event.getDeviceArrayList() != null) {
+            db.deleteAllDevice();
 
-        ArrayList<Devices> devices = event.getDevicesArrayList();
-        for (int i = 0; i < devices.size(); i++) {
-            db.insertDevices(devices.get(i).getId(), devices.get(i).getAndroidId(), devices.get(i).getImageUrl(),
-                    devices.get(i).getName(), devices.get(i).getSnippet());
-        }
-
-        Cursor cursor = db.getAllDevice();
-        ArrayList<Devices> myDevices = new ArrayList<>();
-        if (cursor != null) {
-            while (cursor.moveToNext()) {
-                Devices device = new Devices();
-                device.setId(cursor.getInt(0));
-                device.setAndroidId(cursor.getInt(1));
-                device.setImageUrl(cursor.getString(2));
-                device.setName(cursor.getString(3));
-                device.setSnippet(cursor.getString(4));
-
-                myDevices.add(device);
+            ArrayList<Device> devices = event.getDeviceArrayList();
+            for (int i = 0; i < devices.size(); i++) {
+                db.insertDevices(devices.get(i).getId(), devices.get(i).getAndroidId(), devices.get(i).getImageUrl(),
+                        devices.get(i).getName(), devices.get(i).getSnippet());
             }
-            cursor.close();
+
+            if (ReusableClass.haveNetworkConnection(this)) {
+                if (noData)
+                    setListShown(false, true);
+                MyApplication.addJobInBackground(new GetAllAndroidVersionListJob());
+            } else
+                Toast.makeText(this, R.string.error_internet_connection, Toast.LENGTH_LONG).show();
         }
-
-
-        mAdapter.addAll(myDevices);
     }
 
     public void onEventMainThread(GetDeviceListEvent.Fail event) {
@@ -126,5 +131,87 @@ public class MainActivity extends AppCompatActivity {
                     .setPositiveButton("OK", null)
                     .show();
         }
+    }
+
+    public void onEventMainThread(GetAndroidVersionListEvent.Success event) {
+        setListShown(true, true);
+
+        if (event.getandroidVersionArrayList() != null) {
+            db.deleteAllAndroidVersion();
+
+            ArrayList<AndroidVersion> androidVersions = event.getandroidVersionArrayList();
+            for (int i = 0; i < androidVersions.size(); i++) {
+                db.insertAndroidVersions(androidVersions.get(i).getId(), androidVersions.get(i).getName(), androidVersions.get(i).getVersion(),
+                        androidVersions.get(i).getCodename(), androidVersions.get(i).getTarget(), androidVersions.get(i).getDistribution());
+            }
+
+            addingDataToList();
+        }
+    }
+
+    private void addingDataToList() {
+        Cursor cursor = db.getAllDeviceWithAndroidVersion();
+        ArrayList<AndroidDeviceWithVersion> deviceWithVersionsList = new ArrayList<>();
+        if (cursor != null && cursor.getCount() > 0) {
+            while (cursor.moveToNext()) {
+                AndroidDeviceWithVersion androidDeviceWithVersion = new AndroidDeviceWithVersion();
+                androidDeviceWithVersion.setDeviceId(cursor.getInt(0));
+                androidDeviceWithVersion.setImageUrl(cursor.getString(1));
+                androidDeviceWithVersion.setDeviceName(cursor.getString(2));
+                androidDeviceWithVersion.setSnippet(cursor.getString(3));
+                androidDeviceWithVersion.setAndroidVersionId(cursor.getInt(4));
+                androidDeviceWithVersion.setAndroidVersionName(cursor.getString(5));
+                androidDeviceWithVersion.setVersion(cursor.getString(6));
+
+                deviceWithVersionsList.add(androidDeviceWithVersion);
+            }
+            cursor.close();
+        } else {
+            textViewNoData.setVisibility(View.VISIBLE);
+        }
+        mAdapter.addAll(deviceWithVersionsList);
+    }
+
+    private void addingDataToList(Cursor cursor) {
+        ArrayList<AndroidDeviceWithVersion> deviceWithVersionsList = new ArrayList<>();
+        if (cursor != null && cursor.getCount() > 0) {
+            while (cursor.moveToNext()) {
+                AndroidDeviceWithVersion androidDeviceWithVersion = new AndroidDeviceWithVersion();
+                androidDeviceWithVersion.setDeviceId(cursor.getInt(0));
+                androidDeviceWithVersion.setImageUrl(cursor.getString(1));
+                androidDeviceWithVersion.setDeviceName(cursor.getString(2));
+                androidDeviceWithVersion.setSnippet(cursor.getString(3));
+                androidDeviceWithVersion.setAndroidVersionId(cursor.getInt(4));
+                androidDeviceWithVersion.setAndroidVersionName(cursor.getString(5));
+                androidDeviceWithVersion.setVersion(cursor.getString(6));
+
+                deviceWithVersionsList.add(androidDeviceWithVersion);
+            }
+            cursor.close();
+        } else {
+            textViewNoData.setVisibility(View.VISIBLE);
+        }
+        mAdapter.addAll(deviceWithVersionsList);
+    }
+
+    public void onEventMainThread(GetAndroidVersionListEvent.Fail event) {
+        setListShown(true, true);
+        if (event.getEx() != null) {
+            new AlertDialog.Builder(this)
+                    .setMessage(event.getEx().getMessage())
+                    .setPositiveButton("OK", null)
+                    .show();
+        }
+    }
+
+
+    public void onEventMainThread(PassDeviceEvent.Success event) {
+
+        if (ReusableClass.haveNetworkConnection(this)) {
+            setListShown(false, true);
+            MyApplication.addJobInBackground(new GetAllDeviceListJob());
+        } else
+            Toast.makeText(this, R.string.error_internet_connection, Toast.LENGTH_LONG).show();
+        EventBus.getDefault().removeStickyEvent(event);
     }
 }
